rpi5: add OP-TEE support
OP-TEE support for RPI5 differs from the RPI3 one because on RPI5 we
have BL31 only and there is no way to load additional image into
memory. Only option is to attach OP-TEE image to the end of BL31 so
they can be loaded together by VPU.

So this patch assumes that OP-TEE image is located right after BL31 at
address 0x80000. But we can't boot OP-TEE at this address because VPU
will put kernel right behind OP-TEE image, at 0x200000. It is not
possible to change kernel load address with `kernel_address=` option
in config.txt anymore. This option was moved to "legacy options" ([1])
and is not officially supported anymore. That means that BL31 should
move OP-TEE to an empty memory region. By default BL31 moves it at the
end of 1GB block, but this can be changed with BL31_BASE parameter.

BL31 will pass DTB address to the OP-TEE, so it will be able to
reserve memory for itself and provide correct entries in /firmware
section.

Documentation describes how to attach OP-TEE image to BL31.

[1] https://www.raspberrypi.com/documentation/computers/legacy_config_txt.html#kernel_address

Change-Id: I375ca19e7234053b9d6cee6b644ecc55492d85f2
Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>